Refactor to correctly display stdout from tests
A significant number of changes needed to be made so that the end of test
output is correctly determined when the output does not end with a newline.
Specifically the running algorithm needed to print the length out the output
before it's written out.  That way a subsequent filter can read just those
bytes.

To do so the running algorithm now catches the output into a variable and the
structure filter is written in awk to read out the specified number of bytes.
Changes to the output format propogated out to the other filters, in good ways
I think.  Overall they're simpler, and easier to follow.