Require test scripts to run their tests
This major change to the API should allow EXIT traps in tests. Previously traps
were not allowed because ts used it to launch the test suite.  Experiments
demonstrated that EXIT traps set in tests did nothing as a result.

As an added bonus it should now be possible to source ts without running tests,
in order to gain access to ts_version and other variables.