Require . ts at the end of test scripts
Something needs to execute at the end of test scripts to tell ts to start
running the test suite.  Originally a trap was used such the tests would run on
EXIT.  Currently a special function ts is defined such that 'ts .' runs the
tests.  The problem with the trap is that it makes other exit traps impossible;
the problem with the special function is that it's odd.

Rather than do either of these, simply source ts at the end of the script.
Special protection needs to be provided so that setup/teardown are not
overwritten when present, but defined when they are missing.