Add ability to follow source for test names
The idea here is that in some cases users (I) will want to place common tests
in a shared file and source that into multiple others.

It's a tricky thing pulling the source file from a source command because
whitespace in a file name could be mistaken for an argument break by all text
processing commands and you can't eval until the very end.  Therefore make this
advanced functionality opt-in with a special variable.