Simplify raw output format
Instead of this:

  [file:line] test
  P 1
  x

Use:

  [file:line] test
  P 0

This is simpler, but subtly removes a bit of functionality possible in the
previous system - detection of missing newline at the end of the output.  In a
practical sense this is ok.  As long as the output formats correctly the onus
can be on the user to make the output unambigous (ex use an "inspect" format
instead of a raw format).

At one point preserving the trailing newline was essential to correctly
formatting the output (0d64dde).  I'm not entirely sure what changed, but I'm
unable to reproduce malformatted output with the current code.