Add test/helper to each test file
This makes each test file executable on it's own.  The test_ts_runs_file_in_pwd
test had to be rewritten because, by changing dirs, the helper was no longer
available when the test ran.

In some cases the helper was added to the top of the files, in some cases near
the bottom.  The variation exists because I didn't want to update the line
numbers already existing in some tests.  The order does not matter.