Add note about testing background jobs
I debated making bg job cleanup a default for teardown.  Decided against it
because a) it is easily solved and b) it messes with the idea that teardown
does nothing by default.  There is no super in bash to make it easy to do the
default teardown so I feel it's important that it does nothing - that way users
will not accidentally miss out on some expected behavior.