Sketch out -e flag to exec test in current shell
A key problem with this implementation is that while the setup/teardown does
run, the program will exit with 0, always (because it exits with the status of
the ts_report command, which does not know what the status is of the pipeline
before it).  Moreover, it only works for ./test and not for ts test invocation.