Ignore setup/teardown commands on PATH
Previously if a setup (or presumably a teardown) command existed on PATH then
it would be used in tests that do not create a setup function, resulting in all
kinds of potential breakage.  The reason is that the previous `type` check did
not distinguish between functions and commands.

This adds the `ts_function_exists` command which tests for function existance
in a POSIX-friendly fashion.  Specifically `command -v` returns a full path for
a command, the function name for a function, something about aliases, and no
output if nothing resolves.  That means that if `command -v` for a function
matches the function name then we're good.

See: http://pubs.opengroup.org/onlinepubs/000095399/utilities/command.html