Add reporting if skip file cannot be removed
Note this changes the way skips are detected such that a skip means pass
(exit 0) plus the skip file existing, so that a fail (exit 1) can occur
when the skip file cannot be removed.  If skip exits 1 then there is no
way to tell a true skip from skip-due-to-failure-to-remove.

Moving the removal logic to ts_run_test is likely sound as -e can now
hit that path as well.