Get distribution tests working for non-default shells
Note there were some mistakes in the previous setup.  It was basically
starting at the non-default shell but executing tests as if the default
shell.  In order to install ksh,zsh switch to only testing on the latest
of each distro.  Developers can still check out specific distributions
using the README instructions.