Support zsh getopts
zsh has a non-posix implementatin of getopts such that it behaves quite
differently wrt OPTIND and, as a result, exit status in the old loop.
This new loop relies on whether or not ts_option has been set to detect
the presence of an option (instead of exit status), and directly checks
for '--' rather than relying on OPTIND behavior.

Related to #23